requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
apex=0.1=pypi_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
bottleneck=1.3.5=py37h7deecbd_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2022.12.7=py37h06a4308_0
charset-normalizer=3.0.1=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
et-xmlfile=1.1.0=pypi_0
fftw=3.3.9=h27cfd23_1
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
gdcm=1.1=pypi_0
icu=58.2=he6710b0_3
idna=3.4=pypi_0
imageio=2.25.0=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=4.11.3=py37h06a4308_0
importlib_metadata=4.11.3=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.1=py37h06a4308_0
kaggle=1.5.16=pypi_0
kiwisolver=1.4.4=py37h6a678d5_0
libedit=3.1.20221030=h5eee18b_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libllvm11=11.1.0=h9e868ea_6
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
llvmlite=0.39.1=py37he621ea3_0
matplotlib=3.5.3=pypi_0
matplotlib-base=3.1.3=py37hef1b27d_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
ncurses=6.4=h6a678d5_0
networkx=2.6.3=pypi_0
nibabel=4.0.2=pypi_0
ninja=1.11.1=pypi_0
numba=0.56.4=py37h417a72b_0
numexpr=2.8.4=py37he184ba9_0
numpy=1.21.6=pypi_0
numpy-base=1.21.5=py37ha15fc14_3
opencv-contrib-python=4.7.0.68=pypi_0
opencv-python=4.7.0.68=pypi_0
openpyxl=3.1.2=pypi_0
openssl=1.0.2u=h7b6447c_0
packaging=23.0=pypi_0
pandas=1.3.5=py37h8c16a72_0
pillow=9.4.0=pypi_0
pip=22.3.1=py37h06a4308_0
protobuf=4.21.12=pypi_0
pycocotools=2.0.0=pypi_0
pydicom=2.4.3=pypi_0
pylibjpeg=1.4.0=pypi_0
pylibjpeg-libjpeg=1.3.4=pypi_0
pyparsing=3.0.9=py37h06a4308_0
python=3.7.0=h6e4f718_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python-slugify=8.0.1=pypi_0
pytz=2022.7=py37h06a4308_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=py37h5eee18b_1
readline=7.0=h7b6447c_5
requests=2.28.2=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0.2=py37h51133e4_1
scipy=1.7.3=py37h6c91a56_2
seaborn=0.12.2=py37h06a4308_0
setuptools=65.6.3=py37h06a4308_0
shapely=2.0.1=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.33.0=h62c20be_0
tbb=2021.8.0=hdb19cb5_0
tensorboardx=1.2=pypi_0
text-unidecode=1.3=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2021.11.2=pypi_0
tk=8.6.12=h1ccaba5_0
torch=1.11.0+cu113=pypi_0
torchaudio=0.11.0+cu113=pypi_0
torchvision=0.12.0+cu113=pypi_0
tornado=6.2=py37h5eee18b_0
tqdm=4.64.1=pypi_0
typing-extensions=4.4.0=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.14=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.10=h5eee18b_1
yacs=0.1.8=pypi_0
yaml=0.2.5=h7b6447c_0
zipp=3.11.0=py37h06a4308_0
zlib=1.2.13=h5eee18b_0
